requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-WTF==1.1.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.1
passlib==1.7.4
protobuf==4.21.12
PyJWT==2.7.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
Werkzeug==2.3.6
WTForms==3.0.1
